Move infer steps to zenml.steps.inferrer
