+++ zenml/steps/inferrer/inference_file_writer_step.py
@@ -16,7 +16,7 @@ from typing import Text, List
 
 import apache_beam as beam
 
-from zenml.core.steps.inferrer import BaseInferrer
+from zenml.steps.inferrer import BaseInferrer
 
 
 class FileWriterInferrer(BaseInferrer):
